+++ src/core/warnings/edge/repeatedChrome.ts
@@ -1,0 +1,83 @@
+import type { LayoutBlock, PageWarning } from '../../../types/index.js';
+import { horizontalOverlap } from '../../warningTextOverlap.js';
+
+/** Max vertical gap (in PDF points) between a non-repeated body
+ *  block's bottom and a repeated block's top before we consider the
+ *  two visually mashed together. 6pt is roughly half a body line — at
+ *  this distance the LLM-rendered Markdown joins the lines into one
+ *  paragraph and the footer reads as body text. */
+const CHROME_TOO_CLOSE_GAP_PT = 6;
+
+export function detectBodyNearRepeatedChrome(blocks: LayoutBlock[], out: PageWarning[]): void {
+  // For each non-repeated body block, look at every repeated chrome
+  // block on the page and pick the worst geometric relationship to
+  // report:
+  //
+  //   - **Overlap**: the bboxes vertically intersect. Magnitude is
+  //     the true intersection depth (`min(bodyBottom, chromeBottom)
+  //     - max(bodyTop, chrome.y)`), not `-gap`. The naive `-gap`
+  //     would be wildly off when chrome encroaches on the body's
+  //     top edge from above — e.g. a 40pt header sitting at y=80
+  //     with body at y=100,h=600 overlaps by 20pt, but `-gap`
+  //     (`-(80 - 700) = 620`) would report a 620pt overlap and let
+  //     that header outrank a footer that's barely touching the
+  //     body's bottom.
+  //
+  //   - **Gap**: chrome sits strictly below the body bottom with a
+  //     vertical gap < CHROME_TOO_CLOSE_GAP_PT.
+  //
+  // Overlap always wins over gap (it's a worse readability problem
+  // for an LLM reader), and within each category the worst case
+  // wins — deepest overlap, or smallest gap.
+  for (let i = 0; i < blocks.length; i++) {
+    const body = blocks[i];
+    if (body.repeated) continue;
+    const bodyTop = body.y;
+    const bodyBottom = body.y + body.height;
+    let worstOverlap: { depth: number; index: number } | null = null;
+    let worstGap: { gap: number; index: number } | null = null;
+    for (let j = 0; j < blocks.length; j++) {
+      if (i === j) continue;
+      const chrome = blocks[j];
+      if (!chrome.repeated) continue;
+      // Chrome that lives entirely above the body (a running header
+      // above the first body block) is a different geometric
+      // relationship and isn't what this rule is meant to catch.
+      // Comparing chrome-bottom against body-top lets a header that
+      // dips into the body's top STILL fire (overlap case).
+      const chromeBottom = chrome.y + chrome.height;
+      if (chromeBottom <= bodyTop) continue;
+      if (!horizontalOverlap(body, chrome)) continue;
+      const overlapDepth = Math.min(bodyBottom, chromeBottom) - Math.max(bodyTop, chrome.y);
+      if (overlapDepth > 0) {
+        if (worstOverlap === null || overlapDepth > worstOverlap.depth) {
+          worstOverlap = { depth: overlapDepth, index: j };
+        }
+      } else {
+        const gap = chrome.y - bodyBottom;
+        if (gap >= 0 && gap < CHROME_TOO_CLOSE_GAP_PT) {
+          if (worstGap === null || gap < worstGap.gap) {
+            worstGap = { gap, index: j };
+          }
+        }
+      }
+    }
+    if (worstOverlap !== null) {
+      out.push({
+        code: 'body_near_repeated_chrome',
+        severity: 'warning',
+        message: `body block overlaps a repeated chrome block by ${worstOverlap.depth.toFixed(1)}pt — body text and footer/header are visually colliding`,
+        blockIndex: i,
+        otherBlockIndex: worstOverlap.index,
+      });
+    } else if (worstGap !== null) {
+      out.push({
+        code: 'body_near_repeated_chrome',
+        severity: 'warning',
+        message: `body block ends ${worstGap.gap.toFixed(1)}pt above a repeated chrome block (threshold ${CHROME_TOO_CLOSE_GAP_PT}pt) — body text and footer/header may run together for LLM readers`,
+        blockIndex: i,
+        otherBlockIndex: worstGap.index,
+      });
+    }
+  }
+}
